pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tikka-masalai"
version = "0.0.1"
description = "MLOPS project FIB"
authors = [
  { name = "Team Tikka MasalAI" },
]

classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "accelerate>=0.26.0",
    "dagshub>=0.6.3",
    "dvc==3.63.0",
    "ruamel-yaml-clib==0.2.8",
    "dvc-s3>=3.2.2",
    "evaluate>=0.4.6",
    "huggingface-hub>=0.35.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.6",
    "mlflow>=2,<3",
    "numpy<2.0",
    "pandas>=2.3.2",
    "polars>=1.0.0",
    "pyarrow>=4.0.0,<20.0.0",
    "pytest",
    "python-dotenv",
    "ruff",
    "torch==2.2.2",
    "torchvision==0.17.2",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
    "pre-commit>=4.3.0",
    "deepchecks[vision]>=0.19.1",
    "anywidget>=0.9.18",
    "setuptools==80.8.0",
    "great-expectations>=0.18.22",
    "loguru>=0.7.3",
    "codecarbon>=3.0.7",
    "shap>=0.48.0",
    "opencv-python>=4.11.0.86",
    "python-multipart>=0.0.20",
]
requires-python = ">=3.10,<3.13"  # Temporarily cap <3.13 because torch==2.2.2 has no cp313 wheels

# This makes src/ discoverable as a package
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
required-environments = ["sys_platform == 'darwin' and platform_machine == 'x86_64'"]

[tool.ruff]
line-length = 99
src = ["src", "tests"]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

# Per-file ignores moved into the `lint` section (prevents deprecation warnings)
per-file-ignores = { "tests/**" = ["D101", "D102", "D103"] }

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = "--cov=src --cov-report=html:reports/coverage"

[tool.coverage.run]
omit = ["src/data/validate_data.py", "src/data/gx_context_configuration.py", "src/modeling/train.py"]

[dependency-groups]
backend = [
    "fastapi>=0.118.0",
    "httpx>=0.27.2",
    "matplotlib>=3.10.6",
    "python-multipart>=0.0.20",
    "scipy>=1.15.3",
    "torch>=2.2.2",
    "transformers>=4.57.0",
    "uvicorn>=0.37.0",
]
dev = [
    "black>=25.1.0",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
    "nbqa>=1.0.0",
    "pre-commit>=3.4.0",
    "pytest-cov>=7.0.0",
    "mkdocs-include-markdown-plugin>=6.0.0",
    "mkdocs>=1.6.0",
    "mkdocs-mermaid2-plugin",
]
frontend = [
    "numpy>=1.26.4",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
